Support custom pivots via unity:pivot tile collider
This commit adds support for custom pivots on a per-tile basis.  This
can be done in the collision editor for a tile by using a point
collider and giving it a "class" of "unity:pivot".

I've confirmed it works well for Tile objects placed onto an object
layer, as well as tiles placed on the same tile layer as a player
object.